refactor: move ArtifactRecord to homeboy-lifecycle-contract (joins ArtifactViewerLink)
ArtifactRecord is a pure serde record type describing a recorded run artifact.
Its only nested type, ArtifactViewerLink, already lives in homeboy-lifecycle-
contract::artifact_contract (moved there in a recent cut), so move ArtifactRecord
alongside it rather than leaving the pair split across the crate boundary.

- lifecycle-contract's artifact_contract gains ArtifactRecord (reusing the
  default_artifact_type fn already present there); lib.rs re-exports it.
- homeboy-core re-exports crate::observation::{ArtifactRecord, ArtifactViewerLink}
  from the contract so the ~58 ArtifactRecord call sites are unchanged.
- extension bench/artifact_persistence.rs takes ArtifactRecord from the contract
  directly, reducing its crate::observation edge (keeps only its genuine
  ActiveObservation / finding-records behavior deps).

Reunites the artifact record vocabulary in one contract crate. Continues draining
extension's type edges toward a future homeboy-extension extraction.

Refs #8425
