+++ crates/homeboy-lifecycle-contract/src/lib.rs
@@ -13,8 +13,8 @@ pub mod lifecycle;
 pub mod timeline;
 
 pub use artifact_contract::{
-    ArtifactContract, ArtifactViewerLink, EvidenceContract, ARTIFACT_CONTRACT_SCHEMA,
-    EVIDENCE_CONTRACT_SCHEMA,
+    ArtifactContract, ArtifactRecord, ArtifactViewerLink, EvidenceContract,
+    ARTIFACT_CONTRACT_SCHEMA, EVIDENCE_CONTRACT_SCHEMA,
 };
 pub use lifecycle::{
     LifecycleContract, LifecyclePhaseContract, LifecyclePhaseKind, LifecyclePhaseResult,
